Update information
Built by CA Technologies, Jarvis is another open source alternative to Hubot, specifically designed for working with build messages to make the most of CA’s Flowdock, a chat and collaboration platform. Jarvis works on the JVM, so plugins can be created in Groovy, Java, and Clojure, with support for other JVM languages possible as well such as Kotlin. The downside is extensibility — Jarvis only behaves with Flowdock.