Added AWS ChatBot
It's not a framework, rather a dedicated bot that pushes AWS information into a Slack channel.

Due diligence - I'm not connected to AWS in any way. I'm an IBMer who works with ChatOps and I've got a series of links I'd like to add over the next few days, some will be IBM-y and others (like this) will be 3rd party/generic.